refactor(ui): split recording tag computation out of useInputModes
Move the render-phase analytics tag block (recordingActive / testLabel /
runLogLabel / testRunId / kanaInput assignments and their gate-split and
pristine-runId reasoning) into compute-recording-tags.ts as a plain
function called at the same render-phase spot. No behavior change; the
existing suites pass without a single test edit. Brings useInputModes.ts
from 675 back under the 600-line hook guideline (587).
